+++ src/main/kotlin/app/revanced/patcher/Fingerprint.kt
@@ -3,9 +3,8 @@
 package app.revanced.patcher
 
 import app.revanced.patcher.extensions.InstructionExtensions.instructionsOrNull
-import app.revanced.patcher.patch.*
-import app.revanced.patcher.patch.BytecodePatchContext.LookupMaps.Companion.appendParameters
-import app.revanced.patcher.patch.MethodClassPairs
+import app.revanced.patcher.patch.BytecodePatchContext
+import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.util.proxy.ClassProxy
 import com.android.tools.smali.dexlib2.AccessFlags
 import com.android.tools.smali.dexlib2.Opcode
@@ -16,7 +15,17 @@ import com.android.tools.smali.dexlib2.iface.reference.StringReference
 import com.android.tools.smali.dexlib2.util.MethodUtil
 
 /**
- * A fingerprint.
+ * A fingerprint for a method. A fingerprint is a partial description of a method.
+ * It is used to uniquely match a method by its characteristics.
+ *
+ * An example fingerprint for a public method that takes a single string parameter and returns void:
+ * ```
+ * fingerprint {
+ *    accessFlags(AccessFlags.PUBLIC)
+ *    returns("V")
+ *    parameters("Ljava/lang/String;")
+ * }
+ * ```
  *
  * @param accessFlags The exact access flags using values of [AccessFlags].
  * @param returnType The return type. Compared using [String.startsWith].
@@ -35,16 +44,21 @@ class Fingerprint internal constructor(
     internal val custom: ((method: Method, classDef: ClassDef) -> Boolean)?,
     private val fuzzyPatternScanThreshold: Int,
 ) {
+    @Suppress("ktlint:standard:backing-property-naming")
+    // Backing field needed for lazy initialization.
+    private var _matchOrNull: Match? = null
+
     /**
      * The match for this [Fingerprint]. Null if unmatched.
      */
-    var match: Match? = null
-        private set
+    context(BytecodePatchContext)
+    private val matchOrNull: Match?
+        get() = matchOrNull()
 
     /**
-     * Match using [BytecodePatchContext.LookupMaps].
+     * Match using [BytecodePatchContext.lookupMaps].
      *
-     * Generally faster than the other [match] overloads when there are many methods to check for a match.
+     * Generally faster than the other [matchOrNull] overloads when there are many methods to check for a match.
      *
      * Fingerprints can be optimized for performance:
      * - Slowest: Specify [custom] or [opcodes] and nothing else.
@@ -52,48 +66,50 @@ class Fingerprint internal constructor(
      * - Faster: Specify [accessFlags], [returnType] and [parameters].
      * - Fastest: Specify [strings], with at least one string being an exact (non-partial) match.
      *
-     * @param context The context to create mutable proxies for the matched method and its class.
-     * @return True if a match was found or if the fingerprint is already matched to a method, false otherwise.
+     * @return The [Match] if a match was found or if the fingerprint is already matched to a method, null otherwise.
      */
-    internal fun match(context: BytecodePatchContext): Boolean {
-        val lookupMaps = context.lookupMaps
+    context(BytecodePatchContext)
+    internal fun matchOrNull(): Match? {
+        if (_matchOrNull != null) return _matchOrNull
 
-        fun Fingerprint.match(methodClasses: MethodClassPairs): Boolean {
+        var match = strings?.mapNotNull {
+            lookupMaps.methodsByStrings[it]
+        }?.minByOrNull { it.size }?.let { methodClasses ->
             methodClasses.forEach { (classDef, method) ->
-                if (match(context, classDef, method)) return true
+                val match = matchOrNull(classDef, method)
+                if (match != null) return@let match
             }
-            return false
+
+            null
+        }
+        if (match != null) return match
+
+        classes.forEach { classDef ->
+            match = matchOrNull(classDef)
+            if (match != null) return match
         }
 
-        // TODO: If only one string is necessary, why not use a single string for every fingerprint?
-        if (strings?.firstNotNullOfOrNull { lookupMaps.methodsByStrings[it] }?.let(::match) == true) {
-            return true
-        }
-
-        context.classes.forEach { classDef ->
-            if (match(context, classDef)) return true
-        }
-
-        return false
+        return null
     }
 
     /**
      * Match using a [ClassDef].
      *
      * @param classDef The class to match against.
-     * @param context The context to create mutable proxies for the matched method and its class.
-     * @return True if a match was found or if the fingerprint is already matched to a method, false otherwise.
+     * @return The [Match] if a match was found or if the fingerprint is already matched to a method, null otherwise.
      */
-    fun match(
-        context: BytecodePatchContext,
+    context(BytecodePatchContext)
+    fun matchOrNull(
         classDef: ClassDef,
-    ): Boolean {
+    ): Match? {
+        if (_matchOrNull != null) return _matchOrNull
+
         for (method in classDef.methods) {
-            if (match(context, method, classDef)) {
-                return true
-            }
+            val match = matchOrNull(method, classDef)
+            if (match != null) return match
         }
-        return false
+
+        return null
     }
 
     /**
@@ -101,35 +117,33 @@ class Fingerprint internal constructor(
      * The class is retrieved from the method.
      *
      * @param method The method to match against.
-     * @param context The context to create mutable proxies for the matched method and its class.
-     * @return True if a match was found or if the fingerprint is already matched to a method, false otherwise.
+     * @return The [Match] if a match was found or if the fingerprint is already matched to a method, null otherwise.
      */
-    fun match(
-        context: BytecodePatchContext,
+    context(BytecodePatchContext)
+    fun matchOrNull(
         method: Method,
-    ) = match(context, method, context.classBy { method.definingClass == it.type }!!.immutableClass)
+    ) = matchOrNull(method, classBy { method.definingClass == it.type }!!.immutableClass)
 
     /**
      * Match using a [Method].
      *
      * @param method The method to match against.
      * @param classDef The class the method is a member of.
-     * @param context The context to create mutable proxies for the matched method and its class.
-     * @return True if a match was found or if the fingerprint is already matched to a method, false otherwise.
+     * @return The [Match] if a match was found or if the fingerprint is already matched to a method, null otherwise.
      */
-    internal fun match(
-        context: BytecodePatchContext,
+    context(BytecodePatchContext)
+    fun matchOrNull(
         method: Method,
         classDef: ClassDef,
-    ): Boolean {
-        if (match != null) return true
+    ): Match? {
+        if (_matchOrNull != null) return _matchOrNull
 
         if (returnType != null && !method.returnType.startsWith(returnType)) {
-            return false
+            return null
         }
 
         if (accessFlags != null && accessFlags != method.accessFlags) {
-            return false
+            return null
         }
 
         fun parametersEqual(
@@ -146,17 +160,17 @@ class Fingerprint internal constructor(
 
         // TODO: parseParameters()
         if (parameters != null && !parametersEqual(parameters, method.parameterTypes)) {
-            return false
+            return null
         }
 
         if (custom != null && !custom.invoke(method, classDef)) {
-            return false
+            return null
         }
 
         val stringMatches: List<Match.StringMatch>? =
             if (strings != null) {
                 buildList {
-                    val instructions = method.instructionsOrNull ?: return false
+                    val instructions = method.instructionsOrNull ?: return null
 
                     val stringsList = strings.toMutableList()
 
@@ -176,14 +190,14 @@ class Fingerprint internal constructor(
                         stringsList.removeAt(index)
                     }
 
-                    if (stringsList.isNotEmpty()) return false
+                    if (stringsList.isNotEmpty()) return null
                 }
             } else {
                 null
             }
 
         val patternMatch = if (opcodes != null) {
-            val instructions = method.instructionsOrNull ?: return false
+            val instructions = method.instructionsOrNull ?: return null
 
             fun patternScan(): Match.PatternMatch? {
                 val fingerprintFuzzyPatternScanThreshold = fuzzyPatternScanThreshold
@@ -222,61 +236,217 @@ class Fingerprint internal constructor(
                 return null
             }
 
-            patternScan() ?: return false
+            patternScan() ?: return null
         } else {
             null
         }
 
-        match = Match(
+        _matchOrNull = Match(
             method,
-            classDef,
             patternMatch,
             stringMatches,
-            context,
+            classDef,
         )
 
-        return true
+        return _matchOrNull
     }
+
+    private val exception get() = PatchException("Failed to match the fingerprint: $this")
+
+    /**
+     * The match for this [Fingerprint].
+     *
+     * @throws PatchException If the [Fingerprint] has not been matched.
+     */
+    context(BytecodePatchContext)
+    private val match
+        get() = matchOrNull ?: throw exception
+
+    /**
+     * Match using a [ClassDef].
+     *
+     * @param classDef The class to match against.
+     * @return The [Match] if a match was found or if the fingerprint is already matched to a method, null otherwise.
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    fun match(
+        classDef: ClassDef,
+    ) = matchOrNull(classDef) ?: throw exception
+
+    /**
+     * Match using a [Method].
+     * The class is retrieved from the method.
+     *
+     * @param method The method to match against.
+     * @return The [Match] if a match was found or if the fingerprint is already matched to a method, null otherwise.
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    fun match(
+        method: Method,
+    ) = matchOrNull(method) ?: throw exception
+
+    /**
+     * Match using a [Method].
+     *
+     * @param method The method to match against.
+     * @param classDef The class the method is a member of.
+     * @return The [Match] if a match was found or if the fingerprint is already matched to a method, null otherwise.
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    fun match(
+        method: Method,
+        classDef: ClassDef,
+    ) = matchOrNull(method, classDef) ?: throw exception
+
+    /**
+     * The class the matching method is a member of.
+     */
+    context(BytecodePatchContext)
+    val originalClassDefOrNull
+        get() = matchOrNull?.originalClassDef
+
+    /**
+     * The matching method.
+     */
+    context(BytecodePatchContext)
+    val originalMethodOrNull
+        get() = matchOrNull?.originalMethod
+
+    /**
+     * The mutable version of [originalClassDefOrNull].
+     *
+     * Accessing this property allocates a [ClassProxy].
+     * Use [originalClassDefOrNull] if mutable access is not required.
+     */
+    context(BytecodePatchContext)
+    val classDefOrNull
+        get() = matchOrNull?.classDef
+
+    /**
+     * The mutable version of [originalMethodOrNull].
+     *
+     * Accessing this property allocates a [ClassProxy].
+     * Use [originalMethodOrNull] if mutable access is not required.
+     */
+    context(BytecodePatchContext)
+    val methodOrNull
+        get() = matchOrNull?.method
+
+    /**
+     * The match for the opcode pattern.
+     */
+    context(BytecodePatchContext)
+    val patternMatchOrNull
+        get() = matchOrNull?.patternMatch
+
+    /**
+     * The matches for the strings.
+     */
+    context(BytecodePatchContext)
+    val stringMatchesOrNull
+        get() = matchOrNull?.stringMatches
+
+    /**
+     * The class the matching method is a member of.
+     *
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    val originalClassDef
+        get() = match.originalClassDef
+
+    /**
+     * The matching method.
+     *
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    val originalMethod
+        get() = match.originalMethod
+
+    /**
+     * The mutable version of [originalClassDef].
+     *
+     * Accessing this property allocates a [ClassProxy].
+     * Use [originalClassDef] if mutable access is not required.
+     *
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    val classDef
+        get() = match.classDef
+
+    /**
+     * The mutable version of [originalMethod].
+     *
+     * Accessing this property allocates a [ClassProxy].
+     * Use [originalMethod] if mutable access is not required.
+     *
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    val method
+        get() = match.method
+
+    /**
+     * The match for the opcode pattern.
+     *
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    val patternMatch
+        get() = match.patternMatch
+
+    /**
+     * The matches for the strings.
+     *
+     * @throws PatchException If the fingerprint has not been matched.
+     */
+    context(BytecodePatchContext)
+    val stringMatches
+        get() = match.stringMatches
 }
 
 /**
- * A match for a [Fingerprint].
+ * A match of a [Fingerprint].
  *
- * @param method The matching method.
- * @param classDef The class the matching method is a member of.
+ * @param originalClassDef The class the matching method is a member of.
+ * @param originalMethod The matching method.
  * @param patternMatch The match for the opcode pattern.
  * @param stringMatches The matches for the strings.
- * @param context The context to create mutable proxies in.
  */
-class Match(
-    val method: Method,
-    val classDef: ClassDef,
+context(BytecodePatchContext)
+class Match internal constructor(
+    val originalMethod: Method,
     val patternMatch: PatternMatch?,
     val stringMatches: List<StringMatch>?,
-    internal val context: BytecodePatchContext,
+    val originalClassDef: ClassDef,
 ) {
     /**
-     * The mutable version of [classDef].
+     * The mutable version of [originalClassDef].
      *
      * Accessing this property allocates a [ClassProxy].
-     * Use [classDef] if mutable access is not required.
+     * Use [originalClassDef] if mutable access is not required.
      */
-    val mutableClass by lazy { context.proxy(classDef).mutableClass }
+    val classDef by lazy { proxy(originalClassDef).mutableClass }
 
     /**
-     * The mutable version of [method].
+     * The mutable version of [originalMethod].
      *
      * Accessing this property allocates a [ClassProxy].
-     * Use [method] if mutable access is not required.
+     * Use [originalMethod] if mutable access is not required.
      */
-    val mutableMethod by lazy { mutableClass.methods.first { MethodUtil.methodSignaturesMatch(it, method) } }
+    val method by lazy { classDef.methods.first { MethodUtil.methodSignaturesMatch(it, originalMethod) } }
 
     /**
      * A match for an opcode pattern.
      * @param startIndex The index of the first opcode of the pattern in the method.
      * @param endIndex The index of the last opcode of the pattern in the method.
      */
-    class PatternMatch(
+    class PatternMatch internal constructor(
         val startIndex: Int,
         val endIndex: Int,
     )
@@ -287,7 +457,7 @@ class Match(
      * @param string The string that matched.
      * @param index The index of the instruction in the method.
      */
-    class StringMatch(val string: String, val index: Int)
+    class StringMatch internal constructor(val string: String, val index: Int)
 }
 
 /**
@@ -336,7 +506,7 @@ class FingerprintBuilder internal constructor(
      *
      * @param returnType The return type compared using [String.startsWith].
      */
-    infix fun returns(returnType: String) {
+    fun returns(returnType: String) {
         this.returnType = returnType
     }
 
@@ -427,19 +597,3 @@ fun fingerprint(
     fuzzyPatternScanThreshold: Int = 0,
     block: FingerprintBuilder.() -> Unit,
 ) = FingerprintBuilder(fuzzyPatternScanThreshold).apply(block).build()
-
-/**
- * Create a [Fingerprint] and add it to the set of fingerprints.
- *
- * @param fuzzyPatternScanThreshold The threshold for fuzzy pattern scanning. Default is 0.
- * @param block The block to build the [Fingerprint].
- *
- * @return The created [Fingerprint].
- */
-fun BytecodePatchBuilder.fingerprint(
-    fuzzyPatternScanThreshold: Int = 0,
-    block: FingerprintBuilder.() -> Unit,
-) = app.revanced.patcher.fingerprint(
-    fuzzyPatternScanThreshold,
-    block,
-)() // Invoke to add it.